aaaa/backend/requirements-prod.txt
fastapi==0.110.1
uvicorn[standard]==0.24.0
motor==3.3.1
pydantic==2.12.5
python-dotenv==1.2.2
openai==1.99.9
requests==2.32.5
python-multipart==0.0.22
bcrypt==4.1.3
passlib==1.7.4
python-jose==3.5.0
stripe==7.0.0
pymongo==4.5.0
google-api-python-client>=2.149.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0